:hammer: refactor code for reuse
