+++ components/CustomFooter.vue
@@ -37,26 +37,13 @@
 </template>
 
 <script>
+import {copyToClipboard} from "~/plugins/clipboard"
+
 export default {
   name: 'CustomFooter',
   methods: {
     handleCopy() {
-      var d = document.getElementsByTagName('code')[0].innerHTML
-      d = d.replace(/<\/?span[^>]*>/g, '')
-      navigator.permissions
-        .query({ name: 'clipboard-write' })
-        .then((result) => {
-          if (result.state == 'granted' || result.state == 'prompt') {
-            navigator.clipboard.writeText(d).then(
-              function () {
-                console.log('copied to clipboard')
-              },
-              function () {
-                console.error('Unable to copy to clipboard')
-              }
-            )
-          }
-        })
+      copyToClipboard()
     },
   },
 }
