+++ nuxt.config.js
@@ -21,7 +21,7 @@ export default {
   css: [],
 
   // Plugins to run before rendering page (https://go.nuxtjs.dev/config-plugins)
-  plugins: [],
+  plugins: ['~/plugins/clipboard.js'],
 
   // Auto import components (https://go.nuxtjs.dev/config-components)
   components: true,
